+++ src/app/components/button/Button.tsx
@@ -5,11 +5,21 @@ interface ButtonProps {
   text: string
   onClick: (e: React.FormEvent) => void
   disabled?: boolean
+  variant?: 'primary' | 'secondary'
 }
 
-const Button: React.FC<ButtonProps> = ({ text, onClick, disabled = false }) => {
+const Button: React.FC<ButtonProps> = ({
+  text,
+  onClick,
+  disabled = false,
+  variant = 'primary',
+}) => {
   return (
-    <button className="form-button" onClick={onClick} disabled={disabled}>
+    <button
+      className={`button button--${variant} ${disabled ? 'button--disabled' : ''}`}
+      onClick={onClick}
+      disabled={disabled}
+    >
       {text}
     </button>
   )
