+++ src/app/components/button/button.css
@@ -1,22 +1,42 @@
-.form-button {
+.button {
   padding: 5px 15px;
   font-size: 15px;
   font-weight: bold;
   color: #ffffff;
-  background: #1e0575;
   border: none;
   border-radius: 25px;
   cursor: pointer;
-
+  transition: all 0.2s ease-in-out;
   box-shadow: 0px 5px 15px rgba(0, 0, 0, 0.272);
 }
 
-.form-button:hover {
-  transform: translateY(-1px);
-  box-shadow: 0px 8px 20px rgb(255, 255, 255);
+/* Variante primaria (por defecto) */
+.button--primary {
+  background: #1e0575;
 }
 
-.form-button:active {
+.button--primary:hover {
+  transform: translateY(-1px);
+  box-shadow: 0px 8px 20px rgba(255, 255, 255, 0.5);
+}
+
+.button--primary:active {
   transform: scale(0.95);
   box-shadow: 0px 3px 10px rgba(65, 97, 255, 0.5);
+}
+
+/* Variante secundaria */
+.button--secondary {
+  background: #757575;
+}
+
+.button--secondary:hover {
+  background: #5a5a5a;
+}
+
+/* Botón deshabilitado */
+.button--disabled {
+  background: #b0b0b0;
+  cursor: not-allowed;
+  box-shadow: none;
 }
