+++ src/app/components/header/header.css
@@ -1,24 +1,48 @@
 .header {
+  width: 15%;
+  height: 100vh;
+  position: fixed;
+  top: 0;
+  left: 0;
+  background-color: #333;
+  color: white;
   display: flex;
   flex-direction: column;
-  justify-content: center;
   align-items: center;
-  background-color: white;
-  padding-bottom: 25px;
+  padding: 20px;
+  box-shadow: 2px 0 5px rgba(0, 0, 0, 0.1);
 }
 
-nav {
+.header h1 {
+  margin-bottom: 20px;
+  color: white;
+}
+
+.nav {
   width: 100%;
 }
 
 .link-list {
   list-style-type: none;
   display: flex;
-  justify-content: space-around;
-  margin: 0;
+  flex-direction: column;
+  align-items: start;
   padding: 0;
+}
+
+.link-list li {
+  width: 100%;
 }
 
 .link {
   text-decoration: none;
+  color: white;
+  padding: 10px 0;
+  width: 100%;
+  text-align: left;
+  transition: background-color 0.3s ease;
+}
+
+.link:hover {
+  background-color: #575757;
 }
