+++ src/app/components/pagination/pagination.css
@@ -4,22 +4,34 @@
   margin-top: 20px;
 }
 
-.pagination button {
+.pagination__button {
   padding: 5px 10px;
   border: 1px solid #ccc;
   cursor: pointer;
-  border-radius: 15%;
+  border-radius: 5px;
+  transition: all 0.2s ease-in-out;
 }
 
-.pagination button.active {
+.pagination__button--prev,
+.pagination__button--next {
+  background-color: #f0f0f0;
+}
+
+.pagination__button--active {
   background-color: #0c3f75;
   color: white;
 }
 
-.pagination button:hover {
+.pagination__button:hover {
   transform: translateY(-1px);
 }
-.pagination button:active {
+
+.pagination__button:active {
   transform: scale(0.95);
   box-shadow: 0px 3px 10px rgba(255, 255, 255, 0.5);
+}
+
+.pagination__button:disabled {
+  background-color: #e0e0e0;
+  cursor: not-allowed;
 }
