+++ src/app/components/searchpost/searchpost.css
@@ -1,0 +1,57 @@
+.search-container {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin: 5px;
+}
+
+.search-container__input {
+  width: 100%;
+  max-width: 400px;
+  padding: 10px;
+  border: 1px solid #ccc;
+  border-radius: 5px;
+  font-size: 16px;
+  margin-bottom: 10px;
+  transition: border-color 0.3s ease;
+}
+
+.search-container__input:focus {
+  border-color: #007bff;
+  outline: none;
+}
+
+.search-container__list {
+  position: absolute;
+  top: 100%;
+  width: 100%;
+  max-width: 400px;
+  list-style: none;
+  padding: 0;
+  margin: 0;
+  background-color: #ffffff;
+  border: 1px solid #ccc;
+  border-radius: 5px;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  z-index: 1000;
+}
+
+.search-container__item {
+  padding: 10px;
+  border-bottom: 1px solid #ccc;
+}
+
+.search-container__item:last-child {
+  border-bottom: none;
+}
+
+.search-container__link {
+  text-decoration: none;
+  color: #007bff;
+  transition: color 0.3s ease;
+}
+
+.search-container__link:hover {
+  color: #0056b3;
+}
