+++ src/app/components/subheader/Subheader.tsx
@@ -1,14 +1,9 @@
-import Breadcrumbs from '../breadcrumbs/Breadcrumbs'
 import SearchPost from '../searchpost/SearchPost'
 import './subheader.css'
-import { useLocation } from 'react-router-dom'
 
 function Subheader() {
-  const location = useLocation()
-
   return (
     <div className="subheader">
-      {location.pathname !== '/' && <Breadcrumbs />}
       <SearchPost />
     </div>
   )
