+++ src/app/pages/HomePage.tsx
@@ -1,7 +1,7 @@
 import Header from '../components/header/Header'
-import PostGallery from '../components/PostGallery'
+import PostGallery from '../components/postgallery/PostGallery'
 import Subheader from '../components/subheader/Subheader'
-import '../../styles/home.css'
+import '../../styles/pages/home.css'
 
 function HomePage() {
   return (
@@ -10,7 +10,9 @@ function HomePage() {
       <div className="home-container">
         <Header />
         <div className="home-body-container">
-          <PostGallery numberOfItems={4} />
+          <div className="home__postgallery">
+            <PostGallery numberOfItems={4} />
+          </div>
           <h1>Aquí las gráficas</h1>
         </div>
       </div>
