+++ src/app/pages/PostPage.tsx
@@ -1,5 +1,6 @@
+import Breadcrumbs from '../components/breadcrumbs/Breadcrumbs'
 import Header from '../components/header/Header'
-import PostGallery from '../components/PostGallery'
+import PostGallery from '../components/postgallery/PostGallery'
 import Subheader from '../components/subheader/Subheader'
 
 function PostPage() {
@@ -7,6 +8,7 @@ function PostPage() {
     <div>
       <Header />
       <Subheader />
+      <Breadcrumbs />
       <PostGallery />
     </div>
   )
