+++ src/app/reducers/postReducer.ts
@@ -24,7 +24,6 @@ function postReducer(
     case ADD_POSTS:
       return { posts: [...state.posts, action.payload] }
     case EDIT_POST:
-      console.log('action_payload', action.payload)
       return {
         posts: state.posts.map((post) =>
           post.id === action.payload.id ? action.payload.updatedPost : post
