+++ src/styles/pages/home.css
@@ -1,0 +1,25 @@
+.home-container {
+  display: flex;
+  margin-left: 15%; /* Ajusta el margen para no superponerse con el sidebar */
+  padding: 20px;
+}
+
+.home-body-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  padding: 20px;
+}
+
+.home-body-container h1 {
+  margin-top: 20px;
+}
+
+.home__postgallery {
+  width: 100%;
+}
+
+.home__pagination {
+  margin-top: 20px;
+}
